Boost and TR1 support is no longer required by configure, even with MPI enabled.
Boost now defaults to disabled if you don't specify a path.
It was no longer required if you didn't have MPI enabled,
and now that we use raw MPI for parallel stuff, we no longer
require it in any case.  I left the checks in, just in case
we need it later, but for the time being we should avoid Boost
to minimize mandatory dependencies, and we can add a
HAVE_BOOST define if we want to support optional boost
coding.

I also made a failed TR1 check no longer a fatal configuration
problem; we aren't using any TR1 code, either.