Infrastructure for new eavlOperation developer paradigm and first operation.
So far, this includes a new map operation that can take any number
of inputs and any number of outputs.  No more 1-to-1 or 3-to-3 or 2-to-1
map operation will be needed.

Infrastructure for doing to includes a tuple, a tuple-of-references,
a new dispatch (which, again, is not specific to the number of inputs),
a method to collect a tuple of values from the input arrays, and a new
array indexing class.

I also made use of the new map operation in the unary math mutator.
I added a negate functor so it would build, but disabled the other
unary operations for now.